code/email-classification/backend/requirements.txt
flask
flask-cors
pdfplumber
pytesseract
transformers
torch
google-generativeai
email
openai
